Move APIUser test helper functs to conftest
... and implement as shared pytest fixtures.
